Changed test helper file name. Removed ComponentName parameter. Moved describe block to snapshot tests.
