+++ src/app/components/Footer/index.test.jsx
@@ -1,5 +1,7 @@
 import Footer from './index';
 
-import testHelper from '../../../../src/__test__/test-helper';
+import testHelper from '../../../../src/__test__/testHelper';
 
-testHelper.shouldMatchSnapshot('Footer', Footer);
+describe(`Footer`, () => {
+  testHelper.shouldMatchSnapshot('Footer', Footer);
+});
